Upgrade to Bevy 0.17.3
- Update all workspace dependencies to Bevy 0.17.3 and Rapier 0.32.0
- Migrate Event system to Message system (Event → Message, EventReader → MessageReader, EventWriter → MessageWriter, send → write)
- Update Query API calls (get_single → single, get_single_mut → single_mut)
- Fix #[require] attribute syntax to use component types directly instead of functions
- Replace OrthographicProjection queries with Projection enum
- Update despawn_recursive calls to despawn
- Fix Color API (rgba → srgba)
- Fix WindowResolution to use u32 instead of f32

+++ ui/src/window.rs
@@ -1,8 +1,8 @@
 use bevy::prelude::*;
 use bevy::window::*;
 
-const SCREEN_WIDTH: f32 = 1920.;
-const SCREEN_HEIGHT: f32 = 1080.;
+const SCREEN_WIDTH: u32 = 1920;
+const SCREEN_HEIGHT: u32 = 1080;
 
 pub fn window_plugin() -> WindowPlugin {
     WindowPlugin {
